finnhub_producer/config.py
import os
from dotenv import load_dotenv

load_dotenv()

FINNHUB_API_KEY = os.getenv('FINNHUB_API_KEY')
KAFKA_BOOTSTRAP_SERVERS = os.getenv('KAFKA_BOOTSTRAP_SERVERS', 'kafka:9092')
KAFKA_TOPIC = os.getenv('KAFKA_TOPIC', 'market')

# Schema Registry (for Avro mode)
SCHEMA_REGISTRY_URL = os.getenv('SCHEMA_REGISTRY_URL', 'http://schema-registry:8081')
USE_AVRO = os.getenv('USE_AVRO', 'false').lower() in ('1', 'true', 'yes')

# Symbols to subscribe
SYMBOLS = [
    'AAPL', 'GOOGL', 'MSFT', 'AMZN', 'TSLA',
    'NVDA', 'META', 'NFLX', 'BINANCE:BTCUSDT',
    'BINANCE:ETHUSDT'
]
