finnhub_producer/requirements.txt
websocket-client==1.6.1
kafka-python==2.0.2
python-dotenv==1.0.0
fastavro==1.9.0
requests==2.31.0
prometheus_client==0.16.0
